// src/api/enums/enums.ts
export enum UserType {
    ADMIN = "0",
    USER = "1"
}

export enum Gender {
    MALE = "0",
    FEMALE = "1"
}

export enum ReservationStatus {
    PENDING = "0",
    APPROVED = "1",
    CANCELLED = "2"
}

export enum RoomType {
    SINGLE = "0",
    DOUBLE = "1",
    TRIPLE = "2",
    SUITE = "3"
}